Cleanup time!
Moving towards a pure nodejs implementation of this library. It will be stream based and will take one or more modules as input and output one or more modules, to be concatenated and minified by another library. The task of this project, from here on out, is only to find dependencies recursively.

Got rid of platform specific stuff, libraries (which will be loaded with npm instead) and optimization specific code, which will not be part of this project anymore